+++ projeto-volt/src/main/java/sptech/school/projetovolt/entity/produto/dto/ProdutoConsultaDTO.java
@@ -10,5 +10,5 @@ public class ProdutoConsultaDTO {
     private Double preco;
     private Integer qtdEstoque;
     private String estadoGeral;
-    private Double desconto;
+    private Integer desconto;
 }
