+++ projeto-volt/src/main/java/sptech/school/projetovolt/entity/produto/dto/ProdutoCriacaoDTO.java
@@ -25,5 +25,5 @@ public class ProdutoCriacaoDTO {
     @Size(max = 45)
     private String estadoGeral;
     @PositiveOrZero
-    private Double desconto;
+    private Integer desconto;
 }
